+++ CHANGELOG.md
@@ -12,10 +12,17 @@
 
 ---
 
-## [v1.1.0](https://github.com/foxifyjs/schema/releases/tag/v1.1.0) - *(2018-02-14)*
+## [next](https://github.com/foxifyjs/schema/releases/tag/next) - _(2018-**-**)_
+
+- :boom: All methods are functions now (`.required` => `.required()`)
+- :zap: Added `min` method to object schema type
+- :zap: Added `max` method to object schema type
+- :zap: Added `length` method to object schema type
+
+## [v1.1.0](https://github.com/foxifyjs/schema/releases/tag/v1.1.0) - _(2018-02-14)_
 
 - :star2: Removed `verifications` from dependencies
 
-## [v1.0.0](https://github.com/foxifyjs/schema/releases/tag/v1.0.0) - *(2018-12-27)*
+## [v1.0.0](https://github.com/foxifyjs/schema/releases/tag/v1.0.0) - _(2018-12-27)_
 
 - :tada: First Release